// src/app/types/questionBounds.ts
interface QuestionBounds {
  addition: {
    min_n1: number;
    max_n1: number;
    min_n2: number;
    max_n2: number;
  };
  subtraction: {
    min_n1: number;
    max_n1: number;
    min_n2: number;
    max_n2: number;
  };
  multiplication: {
    min_n1: number;
    max_n1: number;
    min_n2: number;
    max_n2: number;
  };
  division: {
    min_n1: number;
    max_n1: number;
    min_n2: number;
    max_n2: number;
  };
}

export type { QuestionBounds };